requirements.txt
aiofiles==0.8.0
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
chardet==4.0.0
discord==1.7.3
discord-py-interactions==3.0.2
discord.py==1.7.3
dnspython==2.1.0
idna==3.3
multidict==5.2.0
Pillow==9.0.0
py-rolldice==0.4.0
pymongo==4.0.1
regex==2021.11.10
typing_extensions==4.0.1
yarl==1.7.2
